# docs/openapi/quality-control.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quality Control API",
    "version": "2026-07-07",
    "description": "Documentacao inicial dos endpoints criticos do Quality Control. Esta especificacao deve ser atualizada no mesmo PR em que um endpoint novo for criado ou uma regra de permissao for alterada."
  },
  "servers": [
    {
      "url": "/",
      "description": "Mesmo host da aplicacao Next.js"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Sessao web autenticada usada pelas rotas protegidas."
      }
    },
    "schemas": {
...
      "ProjectSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": ["string", "null"] },
          "status": { "type": "string" }
        },
        "additionalProperties": true
      },
...
    }
  },
  "security": [
    {
      "sessionAuth": []
    }
  ],
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "Lista projetos ativos de uma empresa",
        "parameters": [
          {
            "name": "companySlug",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Projetos listados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ProjectSummary" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "companySlug obrigatorio" },
          "401": { "description": "Nao autenticado" },
          "403": { "description": "Empresa fora do escopo" }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Cria projeto na empresa ativa",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["companySlug", "slug", "name"],
                "properties": {
                  "companySlug": { "type": "string" },
                  "slug": { "type": "string" },
                  "name": { "type": "string" },
                  "description": { "type": ["string", "null"] },
                  "color": { "type": ["string", "null"] },
                  "iconKey": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Projeto criado" },
          "400": { "description": "Payload invalido" },
          "401": { "description": "Nao autenticado" },
          "403": { "description": "Sem permissao" },
          "404": { "description": "Empresa nao encontrada" },
          "409": { "description": "Projeto duplicado" }
        }
      }
    },
...
  }
}